+++ wifipumpkin3/core/common/platforms.py
@@ -17,23 +17,22 @@ from glob import glob
 import warnings, json
 from uuid import uuid1
 
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
 
-loggers = {}
-'''http://stackoverflow.com/questions/17035077/python-logging-to-multiple-log-files-from-different-classes'''
-def setup_logger(logger_name, log_file,key=str(), level=logging.INFO):
-    global loggers
-    if loggers.get(logger_name):
-        return loggers.get(logger_name)
-    else:
-        logger = logging.getLogger(logger_name)
-        logger.propagate = False
-        formatter = logging.Formatter('SessionID[{}] %(asctime)s : %(message)s'.format(key))
-        fileHandler = logging.FileHandler(log_file, mode='a')
-        fileHandler.setFormatter(formatter)
-        logger.setLevel(logging.INFO)
-        logger.addHandler(fileHandler)
-    return logger
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
 
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 class Linux(QtCore.QObject):