+++ wifipumpkin3/core/common/uimodel.py
@@ -4,6 +4,23 @@ import wifipumpkin3.core.utility.constants as C
 from wifipumpkin3.core.utility.collection import SettingsINI
 from wifipumpkin3.core.common.platforms import Linux
 
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
+
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 class CoreSettings(Linux):
 
     Name = "General"
@@ -13,7 +30,6 @@ class CoreSettings(Linux):
     Icon=None
     __subitem=False
     conf={}
-
 
     def __init__(self,parent=0,FSettings=None):
         super(CoreSettings,self).__init__()
@@ -31,8 +47,6 @@ class CoreSettings(Linux):
     def osWalkCallback(self,arg,directory,files):
         pass
 
-
-
 class TabsWidget(Qt.QObject):
     Name="Generic"
     ID = "Generic"
@@ -41,15 +55,12 @@ class TabsWidget(Qt.QObject):
     def __init__(self,parent=0,FSettings=None):
         super(TabsWidget,self).__init__()
         self.setObjectName(self.Name)
-        #self.setTitle("{}".format(self.Name))
         self.conf = SuperSettings.getInstance()
         self.parent = parent
 
     @property
     def isSubitem(self):
         return self.__subitem
-
-
 
 class PluginsUI(Qt.QObject):
     Name = "Default"